[#442] Use homemade PDF icon
This replaces the one that had a potential trademark issue with the
Adobe logo in it.

Also contains the hidpi version of the icon that is automatically
picked up in most cases by Eclipse based on the standard naming
scheme of @2x. See https://www.eclipse.org/eclipse/news/4.6/platform_isv.php#high-dpi-icons