Mark all Activators as non-API
This is done with annotations which does not have the "x-friends"
concept that exists for packages. This change causes warnings
because some of the Activators are used by other bundles.

Preferably the Activators should be in a private package and marked
as x-internal and/or x-friends. However I didn't want to restructure
code at this point.

By marking these annotations it means moving the classes later will
not cause API errors and better declares the intention that activators
should not be used by clients.

If any activators do have methods that should be used, those probably
should be moved to new classes that are API and not tied to the
activator (At the API level)