Mark third-party code as x-friends
The embedcdt bundles should not be exporting any packages that are not
embedcdt namespace. It is a simrel requirement that we will have to
fix for 2021-03 to make these bundles separate or (in the case of
semver) change package names.

For now, mark the packages as internal to embedcdt (by friending with
embedcdt bundles). This at least means when the problem is resolved
we don't have API breakages to deal with.

There are some types from these exported third-party packages that
had been part of the API. They have been marked as non-API so they
can be changed without problem too. If they really are intended to be
API, then hopefully the comment will help extenders.

However, this does not prevent wiring errors in OSGi. For example, if
another bundle does Import-Package: org.json.simple they can get
the one provided in the embedcdt which may not be expected, especially
if that pulls in other stuff like CDT. No one has complained
about it in 2020-12 milestones, so we are OK for now, but it should
be fixed ASAP.