Export all packages and mark all internal packages as such
x-internal is used for packages not used elsewhere in embedcdt.
x-friends for those used elsewhere in embedcdt (x-friends means
internal, except to the list of bundles)

x-internal applied to all packages with internal in the name.

It may be worth adding x-internal to additional packages if we
don't want to support them as API. It is not necessary to rename
a package, simply marking it x-internal is sufficient.