Bug #457: Scripts to apply code settings
This is the basis for running all the code cleanups. The sources
here are adapted from CDT's ones of the same names in
releng/scripts/do_all_code_cleanups.sh

The warning/error and formatting settings need to be stored in the project,
the org.eclipse.embedcdt.core's project is the "source" project for these
settings and the initial values for all these settings is from CDT. The
CDT ones are largely bases on default Eclipse ones.

The only change from CDT's settings is that non-externalized
strings are not a warning as embedcdt seems to have too many
non-externalized strings to make this warning useful.

The other significant project setting is that code cleanups are
applied automatically on save, this helps enforce contributors
match embedcdt's code style.

Subsequent commits provide the results of automatically running
all these cleanups.