Bug #459 Add a target file for Embed CDT
Using a target file enables a way to share consistent plug-ins in
the runtime Eclipse instance across developers and ensures that
the target platform that Maven Tycho uses is the same as what is
used in Eclipse.

There is no requirement to use the target platform in the IDE, the
current method of installing all the dependencies in the development
Eclipse instance is ok.