Use new ALL_FLAGS feature of CDT for discovery
CDT 11 includes a new ${ALL_FLAGS} that can be specified in the
scanner discovery to pick up all command line options specified
in the managed build.

See the N&N entry at:

https://github.com/jonah/cdt/blob/main/NewAndNoteworthy/CDT-11.0.md#scanner-discovery-considering-all-flags

This change is not required as CDT's change is non-breaking. But
to use ALL_FLAGS you need CDT 11 or newer, specifically for
org.eclipse.cdt.managedbuilder.core;bundle-version="9.5.0"