Add ability to customize GDB port when auto-starting gdb server
When starting a gdbserver (like openocd) connected to a multi-core
SoC the gdb server may open multiple ports. This change allows
user to configure which port GDB connect to rather than forcing
GDB to connect to the first port.

When the ports do not match a warning triangle is displayed which
can be clicked on to restore the matching values.

When a user changes the "GDB port" in the "Start TYPE locally"
section that is reflected in the "Port number" in the "Remote Target"
section.

Changes to "Port number" in the "Remote Target" which lead to the ports
not matching will display the warning triangle.

Fixes #569