Add some visual separation between toolchain settings
- groups around the global and per-toolchain settings with labels
- whitespace between different toolchain settings
- new logic that shows which projects + build configs are using
each displayed toolchain
- added links on the project properties pages back to the global
and workspace preferences
- updated the descriptions in the pages

The change in XpackDirectoryNotStrictFieldEditor fixes the
implementation to allow doFillIntoGrid to be called multiple
times. This is how the other field editors handle it. This
is needed now because of how the layout is now being handled
with the extra layer of UI hierarchy.

Fixes #574