Rework setting PATH variable to not store in a File temporarily
The old code placed multiple entries for adding to PATH in a File
temporarily to get the absolute path. This worked generally on
non-Windows platforms because often there was only a single entry.

However on Windows there are generally two entries. Adding to that
is the UNC path names that are sometimes present on Windows and
the temporary use of File causes corruption in the settings.

Instead collect the new PATH entries in a list of strings, and then
before making the final PATH setting, run each individual entry
through new File(entry).getAbsolutePath()

Fixes #230