Enable Externalize Strings tooling
In commit 485c41b6f59a327d2aae3a890e7f2b90f56ef037 the code was refactored
in a such a way that prevented the externalize string tooling in
Eclipse from working properly.

This change provides the messages classes with a way that the tooling
in Eclipse works, enabling Ctrl-Click and hovering on the externalized
strings to work as expected and allows the externalize strings wizard
to find the correct Messages class to add new strings to.