Use more modern NLS mechanism
This change uses consistent and modern NLS mechanism. This makes it
easier to handle refactoring code and applying the same change to
multiple classes like recently done in #575

Part 1:

In commit 485c41b6f59a327d2aae3a890e7f2b90f56ef037 the code was refactored
in a such a way that prevented the externalize string tooling in
Eclipse from working properly.

This change provides the messages classes with a way that the tooling
in Eclipse works, enabling Ctrl-Click and hovering on the externalized
strings to work as expected and allows the externalize strings wizard
to find the correct Messages class to add new strings to.

Part 2:

Using fields instead of getString provides compile
time verification that the key is correct. This is extra helpful
when copying code around as it is immediately apparent that the
a message is missing from the Messages.java class.

In fact when I refactored the code to use the static way I found
a handful of messages that were missing.

Disadvantage:

There is one downside to the above changes. All the message keys that
were not already Java identifiers needed modifying. For example,
DebuggerTab.doStartGdbServer_Text needed changing to
DebuggerTab_doStartGdbServer_Text. For embed-cdt itself this
isn't a problem. But for any translations that have been
done for messages.properties they will have to update the keys
as well to change the . to _

Other advantage:

The other advantage is if a entry is missing from the messages.properties
class, it is logged when any message in the Messages class is loaded
rather than when the missing one is loaded. A missing entry shows up
like this:

```
Warning: NLS missing message: ProjectToolchainsPathsPropertiesPage_description in: org.eclipse.embedcdt.internal.managedbuild.cross.arm.ui.messages
```

This significantly increases the changes of spotting missing messages
during development because not every single code path needs to be
visited.